Refactor tests of manage_views
